Start to split up the json_frontend
