+++ json_frontend/src/json_lang.rs
@@ -1,0 +1,74 @@
+use serde::Deserialize;
+use serde_json::Value;
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum JSONLang {
+    Module { name: String, decls: Vec<Decl> },
+}
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Decl {
+    FwdDecl {
+        name: String,
+        visibility: Visibility,
+        r#type: Type,
+        args: FuncArgs,
+    },
+    FuncDecl {
+        name: String,
+        visibility: Visibility,
+        r#type: Type,
+        args: FuncArgs,
+        stmts: Vec<Stmt>,
+    },
+}
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum FuncArgs {
+    Fixed(Vec<FuncArg>),
+    Variadic(FuncArg, Vec<FuncArg>),
+}
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum FuncArg {
+    Named { name: String, r#type: Type },
+}
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Stmt {
+    Ret { value: Expr },
+    VarDecl { name: String, value: Expr },
+}
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Expr {
+    Const {
+        value: Value,
+        r#type: Type,
+    },
+    FuncCall {
+        name: String,
+        r#type: Type,
+        args: Vec<Expr>,
+    },
+}
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Visibility {
+    Public,
+    Private,
+}
+
+#[derive(Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Type {
+    Int32,
+    Str,
+}
